+++ CHANGES.rst
@@ -12,7 +12,8 @@ UNRELEASED
 - Add support for Python 3.9, 3.10, 3.11 and 3.12.
 - Drop support for Python 3.8 or older.
 - Add support for WTForms 3.2. The minimum supported WTForms version is now 3.1.0.
-- Remove `six` dependency.
+- Remove `email_validator` and `six` dependencies.
+- Remove `_compat` module.
 
 
 0.10.5 (2021-01-17)